Working on language packs
Did a session with a lot of debugging and trying to get multiple properties file to work together in one class. Added a CustomProperties class to include a partial path name that contains the most recent sub-folder and the type-name: i.e. CustomProperties uiText = new CustomProperties("uitext/UIText");. The rest of the prefix is appended as before. TODO: Documente the whole GameMenu class.